refactor: Cleanup kitscreen
- KitScreen doesn't need to know anything except its high level components. Previously kitItems was being passed down as props. But that responsibility should go to OrbManager
